Actively rerender when the stage is dirty
Actively detect if the stage is dirty and rerender the stage. It prevents the shadows from being generated when the scripts are dragged over the stage and some unknown reasons. But your hack method is still important because the reason of  how shadows of the only sprite come from the interactive events is still unclear and the hack method does work in these situations.